bds_b3I: trying to fix bug with b1i pvt

+++ src/algorithms/tracking/gnuradio_blocks/dll_pll_veml_tracking.cc
@@ -42,7 +42,6 @@
 #include "GPS_L5.h"
 #include "Galileo_E1.h"
 #include "Galileo_E5a.h"
-#include "Beidou_B1I.h"
 #include "MATH_CONSTANTS.h"
 #include "beidou_b3i_signal_processing.h"
 #include "beidou_b1i_signal_processing.h"
@@ -288,14 +287,14 @@ dll_pll_veml_tracking::dll_pll_veml_tracking(const Dll_Pll_Conf &conf_) : gr::bl
                     d_signal_carrier_freq = BEIDOU_B1I_FREQ_HZ;
                     d_code_period = BEIDOU_B1I_CODE_PERIOD;
                     d_code_chip_rate = BEIDOU_B1I_CODE_RATE_HZ;
-                    d_code_length_chips = static_cast<unsigned int>(BEIDOU_B1I_CODE_LENGTH_CHIPS);
+                    d_code_length_chips = static_cast<uint32_t>(BEIDOU_B1I_CODE_LENGTH_CHIPS);
                     d_symbols_per_bit = BEIDOU_B1I_TELEMETRY_SYMBOLS_PER_BIT;
                     d_correlation_length_ms = 1;
                     d_code_samples_per_chip = 1;
                     d_secondary = true;
                     trk_parameters.track_pilot = false;
                     interchange_iq = false;
-                    d_secondary_code_length = static_cast<unsigned int>(BEIDOU_B1I_SECONDARY_CODE_LENGTH);
+                    d_secondary_code_length = static_cast<uint32_t>(BEIDOU_B1I_SECONDARY_CODE_LENGTH);
                     d_secondary_code_string = const_cast<std::string *>(&BEIDOU_B1I_SECONDARY_CODE_STR);
                 }
             else if (signal_type == "B3")
@@ -304,14 +303,14 @@ dll_pll_veml_tracking::dll_pll_veml_tracking(const Dll_Pll_Conf &conf_) : gr::bl
         	    d_signal_carrier_freq = BEIDOU_B3I_FREQ_HZ;
 				d_code_period = BEIDOU_B3I_CODE_PERIOD;
 				d_code_chip_rate = BEIDOU_B3I_CODE_RATE_HZ;
-				d_code_length_chips = static_cast<unsigned int>(BEIDOU_B3I_CODE_LENGTH_CHIPS);
+				d_code_length_chips = static_cast<uint32_t>(BEIDOU_B3I_CODE_LENGTH_CHIPS);
 				d_symbols_per_bit = BEIDOU_B3I_TELEMETRY_SYMBOLS_PER_BIT;
 				d_correlation_length_ms = 1;
 				d_code_samples_per_chip = 1;
 				d_secondary = true;
 				trk_parameters.track_pilot = false;
 				interchange_iq = false;
-				d_secondary_code_length = static_cast<unsigned int>(BEIDOU_B3I_SECONDARY_CODE_LENGTH);
+				d_secondary_code_length = static_cast<uint32_t>(BEIDOU_B3I_SECONDARY_CODE_LENGTH);
 				d_secondary_code_string = const_cast<std::string *>(&BEIDOU_B3I_SECONDARY_CODE_STR);
                 }
             else
@@ -612,8 +611,8 @@ void dll_pll_veml_tracking::start_tracking()
                     d_preamble_length_symbols = 22;
                     d_preambles_symbols = static_cast<int32_t *>(volk_gnsssdr_malloc(22 * sizeof(int32_t), volk_gnsssdr_get_alignment()));
                     int32_t n = 0;
-                    uint16_t preambles_bits[BEIDOU_B1I_PREAMBLE_LENGTH_BITS] = {1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0};
-					for (uint16_t preambles_bit : preambles_bits)
+                    uint32_t preambles_bits[BEIDOU_B1I_PREAMBLE_LENGTH_BITS] = {1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 0};
+					for (uint32_t preambles_bit : preambles_bits)
 						{
 							for (int32_t j = 0; j < d_symbols_per_bit; j++)
 								{
@@ -653,8 +652,8 @@ void dll_pll_veml_tracking::start_tracking()
 					d_preamble_length_symbols = 22;
 					d_preambles_symbols = static_cast<int32_t *>(volk_gnsssdr_malloc(22 * sizeof(int32_t), volk_gnsssdr_get_alignment()));
 					int32_t n = 0;
-					uint16_t preambles_bits[BEIDOU_B3I_PREAMBLE_LENGTH_BITS] = {1,1,1,0,0,0,1,0,0,1,0};
-                    for (uint16_t preambles_bit : preambles_bits)
+					uint32_t preambles_bits[BEIDOU_B3I_PREAMBLE_LENGTH_BITS] = {1,1,1,0,0,0,1,0,0,1,0};
+                    for (uint32_t preambles_bit : preambles_bits)
                         {
                             for (int32_t j = 0; j < d_symbols_per_bit; j++)
                                 {
